requirements.txt
 flask== 0.12.3 
 flask-mail==0.7.6
 SQLAlchemy>=1.3.0
 flask-sqlalchemy==0.16
 sqlalchemy-migrate==0.7.2
 flask-whooshalchemy==0.55a
 flask-wtf==0.8.4
 pytz==2013b
 flask-babel==0.8
 flup
 pydocumentdb>=1.0.0
 azure-cosmos==3.0.2